Create README.md
Objective: The main objective of the bank database that is taken from the KDD Nuggets referenced data is ensuring seamless transactions of the bank's operations, by taking the structure of the bank as a prime focus element, managing the client interactions with the system effectively, and efficient manner of handling the data.

Within the selected dataset, multiple discrete datasets have been found, each pertaining to a particular aspect of financial activities and transactions. The combination of these separate datasets is what we call the "bank dataset." We have designed a comprehensive and interconnected database system, dubbed the "bank database," to handle and analyze the vast amount of information found in these datasets.

To implement our strategy, we first create unique entities in the database, each of which corresponds to a different dataset. These entities act as well-organized data containers, guaranteeing the appropriate preservation of the relationships and underlying structure among the various forms of banking information. By doing this, we can make the most of this large dataset and realize its full potential, making it more accessible, organized, and insightful.

This structured methodology helps us get valuable insights from the bank dataset and improves the data management process overall. It enables us to monitor financial movements, examine transactional trends, examine client behavior, and offer insightful information to assist strategic decision-making in the financial industry. In essence, our design and implementation of the bank database facilitate a more profound exploration of the complex financial landscape encapsulated by these multiple datasets.